RE-344 Split Up BetterLock Events to its own events
