+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/events/protect/BlockPistonProtect.java
@@ -1,0 +1,40 @@
+package de.relluem94.minecraft.server.spigot.essentials.events.protect;
+
+import de.relluem94.minecraft.server.spigot.essentials.RelluEssentials;
+import de.relluem94.minecraft.server.spigot.essentials.helpers.pojo.ProtectionEntry;
+import org.bukkit.Material;
+import org.bukkit.block.Block;
+import org.bukkit.block.BlockFace;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockPistonExtendEvent;
+import org.bukkit.event.block.BlockPistonRetractEvent;
+import org.jetbrains.annotations.NotNull;
+
+public class BlockPistonProtect implements Listener {
+    @EventHandler
+    public void onBlockPistonExtend(@NotNull BlockPistonExtendEvent e) {
+        for (Block b : e.getBlocks()) {
+            ProtectionEntry protection = RelluEssentials.getInstance().getProtectionAPI().getProtectionEntry(b.getLocation());
+            if (protection != null || isProtected(b, BlockFace.UP) || isProtected(b, BlockFace.DOWN)) {
+                e.setCancelled(!b.getType().equals(Material.WATER));
+                break;
+            }
+        }
+    }
+
+    @EventHandler
+    public void onBlockPistonRetract(@NotNull BlockPistonRetractEvent e) {
+        for (Block b : e.getBlocks()) {
+            ProtectionEntry protection = RelluEssentials.getInstance().getProtectionAPI().getProtectionEntry(b.getLocation());
+            if (protection != null || isProtected(b, BlockFace.UP) || isProtected(b, BlockFace.DOWN)) {
+                e.setCancelled(true);
+                break;
+            }
+        }
+    }
+
+    private boolean isProtected(@NotNull Block b, BlockFace bf){
+        return RelluEssentials.getInstance().getProtectionAPI().getProtectionEntry(b.getRelative(bf).getLocation()) != null;
+    }
+}
