+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/events/protect/BlockRedstoneProtect.java
@@ -1,0 +1,34 @@
+package de.relluem94.minecraft.server.spigot.essentials.events.protect;
+
+import de.relluem94.minecraft.server.spigot.essentials.RelluEssentials;
+import de.relluem94.minecraft.server.spigot.essentials.constants.ProtectionFlags;
+import de.relluem94.minecraft.server.spigot.essentials.helpers.ProtectionHelper;
+import de.relluem94.minecraft.server.spigot.essentials.helpers.pojo.ProtectionEntry;
+import org.bukkit.Location;
+import org.bukkit.block.Block;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockRedstoneEvent;
+import org.jetbrains.annotations.NotNull;
+import org.json.JSONObject;
+
+import static de.relluem94.minecraft.server.spigot.essentials.constants.Constants.PLUGIN_EVENT_PROTECT_FLAGS;
+
+public class BlockRedstoneProtect implements Listener {
+    @EventHandler
+    public void onBlockRedstoneChange(@NotNull BlockRedstoneEvent e) {
+        Block b = e.getBlock();
+        ProtectionEntry protection;
+
+        Location l = ProtectionHelper.getLocationFromBlockAlternateForDoor(b);
+        protection = RelluEssentials.getInstance().getProtectionAPI().getProtectionEntry(l);
+
+        if (protection != null){
+            JSONObject flags = protection.getFlags();
+            boolean isAllowed = (!flags.isEmpty() && flags.has(PLUGIN_EVENT_PROTECT_FLAGS) && flags.getJSONArray(PLUGIN_EVENT_PROTECT_FLAGS).toList().contains(ProtectionFlags.ALLOW_REDSTONE.name()));
+            if(!isAllowed){
+                e.setNewCurrent(e.getOldCurrent());
+            }
+        }
+    }
+}
