+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/helpers/ProtectionHelper.java
@@ -170,4 +170,7 @@ public class ProtectionHelper {
         }
         return location;
     }
+
+
+
 }
